Separate files for contact sheet and rotation curve
Made a new file - "...-rotationcurve" which outputs only the images of the density components and the rotation curves for a singe solution. Output is multiple png files that could be used individually for presentations.
Renamed the previous file "...-rotcurves-contactsheet" and made some minor changes to it. This file creates the contact sheet composed of all the solutions.